修复缩放抖动问题

+++ XImageView/src/main/java/cn/kejin/android/views/XImageView.java
@@ -465,7 +465,7 @@ public class XImageView extends View
 
             float factor = detector.getScaleFactor();
             mBitmapManager.scaleShowBitmap(
-                    (int) detector.getFocusX(), (int) detector.getFocusY(), factor);
+                    detector.getFocusX(), detector.getFocusY(), factor);
 
             return true;
         }
